CAMBIO LAS SALIDAS

+++ Metrics.py
@@ -3,19 +3,19 @@ import numpy as np
 
 def MAE(y_true: np.ndarray, y_pred: np.ndarray) -> np.float64:
     """ Implementation of the Mean Absolute Error (MAE) """
-    return np.mean(np.abs(y_true - y_pred))
+    return np.float64(np.mean(np.abs(y_true - y_pred)))
 
 def MSE(y_true: np.ndarray, y_pred: np.ndarray) -> np.float64:
     """ Implementation of the Mean Squared Error (MSE) """
-    return np.mean((y_true - y_pred)**2)
+    return np.float64(np.mean((y_true - y_pred)**2))
 
 def R2(y_true: np.ndarray, y_pred: np.ndarray) -> np.float64:
     """ Implementation of the R2 metric """
     ss_res = np.sum((y_true - y_pred)**2)
     ss_tot = np.sum((y_true - np.mean(y_true))**2)
-    return 1 - (ss_res/ss_tot)
+    return np.float64(1 - (ss_res/ss_tot))
 
 def Corr(y_true: np.ndarray, y_pred: np.ndarray) -> np.float64:
     """ Implementation of the Pearson's Correlation Coefficient """
     corr = np.corrcoef(y_true, y_pred)
-    return corr[0, 1]
+    return np.float64(corr[0, 1])
